+++ build.sh
@@ -76,6 +76,9 @@ fi
 echo "Syncing version..."
 node .scripts/sync-version.js
 
+echo "Formatting JSON..."
+node .scripts/format-json.js
+
 # Update EGO published version badge
 echo "Updating EGO published version badge..."
 node .scripts/fetch-ego-version.js